Add a new style based on solarized
Add $monospaceFont and $abbvFont in settings so that fonts for different
text can be customized in themes. A new theme called solarized-stylish
is created with adapting fonts with Google Fonts. A minor tweeking of
the solarized-stylish theme is the font size for heading, which have to
be shrinked a bit due to the the font used.